Create NovaProcessor.java
